Subo version andando localmente, con informe, cpplint ready, y con correciones contestadas

+++ common_game_result.cpp
@@ -1,12 +1,12 @@
 #include "common_game_result.h"
 #include <sstream>
 
-GameResult::GameResult(const size_t game_size) : good(0), regular(0),
-												  game_size(game_size) {}
+GameResult::GameResult(const unsigned int game_size) : good(0), regular(0),
+												       game_size(game_size) {}
 
 GameResult::~GameResult() {}
 
-void GameResult::setter(const size_t good, const size_t regular) {
+void GameResult::setter(const unsigned int good, const unsigned int regular) {
 	this->good = good;
 	this->regular = regular;
 }
